worker logic
still remains to fix
